+++ src/gromo/modules/conv2d_growing_module.py
@@ -59,7 +59,7 @@ class Conv2dGrowingModule(GrowingModule):
         stride: int | tuple[int, int] = 1,
         padding: int | tuple[int, int] = 0,
         dilation: int | tuple[int, int] = 1,
-        input_size: tuple[int, int] = (-1, -1),
+        input_size: tuple[int, int] | None = None,
         # groups: int = 1,
         use_bias: bool = True,
         post_layer_function: torch.nn.Module = torch.nn.Identity(),
@@ -102,9 +102,10 @@ class Conv2dGrowingModule(GrowingModule):
             device=device,
             name=name,
         )
+        self.layer: torch.nn.Conv2d
         self.kernel_size = self.layer.kernel_size
 
-        self.input_size: tuple[int, int] = input_size
+        self._input_size: tuple[int, int] | None = input_size
         self.use_bias = use_bias
 
     # Information functions
@@ -456,7 +457,12 @@ class Conv2dGrowingModule(GrowingModule):
                     f"yet for {type(self.previous_module)} as previous module."
                 )
 
-    def update_input_size(self, input_size: tuple[int, int] | None = None) -> None:
+    def update_input_size(
+        self,
+        input_size: tuple[int, int] | torch.Size | None = None,
+        compute_from_previous: bool = False,
+        force_update: bool = True,
+    ) -> tuple[int, int] | None:
         """
         Update the input size of the layer. Either according to the parameter or the input currently stored.
 
@@ -464,25 +470,54 @@ class Conv2dGrowingModule(GrowingModule):
         ----------
         input_size: tuple[int, int] | None
             new input size
+        compute_from_previous: bool
+            whether to compute the input size from the previous module
+            assuming its output size won't be affected by the post-layer function
+        force_update: bool
+            whether to force the update even if the input size is already set
+            (_input_size is not None)
+
+        Returns
+        -------
+        tuple[int, int] | None
+            updated input size if it could be computed, None otherwise
         """
         if input_size is not None:
-            new_size = input_size
+            new_size = tuple(input_size)
         elif self.store_input and self.input is not None:
-            new_size = self.input.shape[-2:]
-        elif self.previous_module and self.previous_module.input_size != (-1, -1):
+            new_size: tuple[int, ...] = tuple(self.input.shape[2:])
+        elif not force_update and self._input_size is not None:
+            return self._input_size
+        elif (
+            compute_from_previous
+            and self.previous_module
+            and (
+                prev_input_size := self.previous_module.update_input_size(
+                    force_update=False
+                )
+            )
+        ):
+            # we get it this way instead of self.previous_module.input_size
+            # to avoid errors if the previous module input size can't be computed
             new_size = compute_output_shape_conv(
-                self.previous_module.input_size, self.previous_module.layer
+                prev_input_size, self.previous_module.layer
             )
         else:
-            raise AssertionError(f"Unable to compute the input size for {self.name}.")
+            # if we cannot compute it, just return the current value
+            return self._input_size
 
-        if self.input_size != (-1, -1) and new_size != self.input_size:
+        if self._input_size is not None and new_size != self._input_size:
             warn(
-                f"The input size of the layer {self.name} has changed from {self.input_size} to {new_size}."
+                f"The input size of the layer {self.name} has changed from {self._input_size} to {new_size}."
                 f"This may lead to errors if the size of the tensor statistics "
                 f"and of the mask tensor T are not updated."
             )
-        self.input_size = new_size
+
+        assert (
+            len(new_size) == 2
+        ), f"The input size should be a tuple of two integers, but got {new_size=}."
+        self._input_size = new_size
+        return self._input_size
 
     def update_computation(self) -> None:
         """
@@ -505,7 +540,7 @@ class RestrictedConv2dGrowingModule(Conv2dGrowingModule):
         stride: int | tuple[int, int] = 1,
         padding: int | tuple[int, int] = 0,
         dilation: int | tuple[int, int] = 1,
-        input_size: tuple[int, int] = (-1, -1),
+        input_size: tuple[int, int] | None = None,
         # groups: int = 1,
         use_bias: bool = True,
         post_layer_function: torch.nn.Module = torch.nn.Identity(),
@@ -901,7 +936,7 @@ class FullConv2dGrowingModule(Conv2dGrowingModule):
         stride: int | tuple[int, int] = 1,
         padding: int | tuple[int, int] = 0,
         dilation: int | tuple[int, int] = 1,
-        input_size: tuple[int, int] = (-1, -1),
+        input_size: tuple[int, int] | None = None,
         # groups: int = 1,
         use_bias: bool = True,
         post_layer_function: torch.nn.Module = torch.nn.Identity(),
@@ -947,10 +982,6 @@ class FullConv2dGrowingModule(Conv2dGrowingModule):
         torch.Tensor
             mask tensor T
         """
-        assert self.input_size != (-1, -1), (
-            f"The input size should be set before computing the mask tensor T "
-            f"for {self.name}."
-        )
         self.layer: torch.nn.Conv2d  # CHECK: why do we need to specify the type here?
         if self._mask_tensor_t is None:
             self._mask_tensor_t = compute_mask_tensor_t(self.input_size, self.layer).to(