+++ tests/test_conv2d_growing_module.py
@@ -363,25 +363,42 @@ class TestConv2dGrowingModule(TorchTestCase):
         self.assertLess(loss.item(), 1e-3)
 
     @unittest_parametrize(({"bias": True}, {"bias": False}))
-    def test_update_input_size(self, bias: bool):
+    def test_input_size(self, bias: bool):
         demo_layer = self.bias_demos[bias]
 
         # error
-        with self.assertRaises(AssertionError):
-            demo_layer.update_input_size()
+        with self.assertRaises(ValueError):
+            demo_layer.input_size
+
+        # no error but get None
+        self.assertIsNone(demo_layer.update_input_size(force_update=False))
 
         # automatic setting
         demo_layer.store_input = True
         demo_layer(self.input_x)
 
-        demo_layer.update_input_size()
         self.assertEqual(demo_layer.input_size, (10, 10))
 
         # manual
         with self.assertWarns(Warning):
-            demo_layer.update_input_size((7, 7))
+            demo_layer.input_size = (7, 7)
 
         self.assertEqual(demo_layer.input_size, (7, 7))
+
+        demo_layer.input_size = None
+        self.assertIsNone(demo_layer._input_size)
+
+    def test_input_size_with_recursive_calls(self):
+        demo_couple: tuple[Conv2dGrowingModule, Conv2dGrowingModule] = self.demo_couple[
+            False
+        ]
+        demo_in, demo_out = demo_couple
+
+        demo_in.input_size = self.input_x.shape[2:]
+        y = demo_in(self.input_x)
+        self.assertEqual(demo_in.input_size, self.input_x.shape[2:])
+        demo_out.update_input_size(compute_from_previous=True)
+        self.assertEqual(demo_out.input_size, y.shape[2:])
 
 
 class TestFullConv2dGrowingModule(TestConv2dGrowingModule):
@@ -473,7 +490,7 @@ class TestFullConv2dGrowingModule(TestConv2dGrowingModule):
         )
 
     def test_mask_tensor_t(self):
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(ValueError):
             _ = self.demo.mask_tensor_t
 
         hin, win = 11, 13
@@ -520,7 +537,7 @@ class TestFullConv2dGrowingModule(TestConv2dGrowingModule):
                 loss.backward()
 
                 demo_couple[0].update_input_size()
-                demo_couple[1].update_input_size()
+                demo_couple[1].update_input_size(compute_from_previous=True)
                 demo_couple[1].tensor_m_prev.update()
 
                 self.assertEqual(